Move DMIDValue and DMIDMessage into own file in DMID directory
