+++ arangod/Pregel/Algos/DMID/DMID.h
@@ -25,6 +25,8 @@
 
 #include "Pregel/Algorithm.h"
 #include "Pregel/CommonFormats.h"
+#include "Pregel/Algos/DMID/DMIDValue.h"
+#include "Pregel/Algos/DMID/DMIDMessage.h"
 
 namespace arangodb {
 namespace pregel {
